requirements.txt
absl-py==2.0.0
altgraph==0.17.2
astunparse==1.6.3
atomicwrites==1.4.1
attrs==22.1.0
Brotli==1.0.9
cachetools==5.3.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cycler==0.11.0
distlib==0.3.5
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.0.2
Flask-Compress==1.10.1
Flask-SQLAlchemy==2.5.1
flatbuffers==23.5.26
fonttools==4.29.1
future==0.18.2
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.58.0
gunicorn==20.1.0
h5py==3.9.0
ibapi @ file:///D:/TWS%20API/source/pythonclient/dist/ibapi-9.76.1-py3-none-any.whl
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.0.1
matplotlib==3.5.1
mplfinance==0.12.8b9
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
passlib==1.7.4
pefile==2022.5.30
Pillow==9.0.1
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.19.6
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
pyparsing==3.0.7
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.7
pywin32-ctypes==0.2.0
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.29
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==1.26.7
virtualenv==20.16.3
Werkzeug==2.0.2
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0